Show how to use Server Actions
