+++ src/app/products/services/mutations/deleteProduct.ts
@@ -1,0 +1,13 @@
+"use server"
+import { db } from "@/db/drizzle-client";
+import { products } from "@/db/schema";
+import { eq } from "drizzle-orm";
+import { revalidatePath } from "next/cache";
+
+// TODO: All these DB editing functions do not have zod validation yet. Should be added.
+export async function deleteProduct(productId: string) {
+  await db
+      .delete(products)
+      .where(eq(products.id, productId as unknown as number));
+  revalidatePath('/products')
+}
